+++ leptos_dom/src/components/errors.rs
@@ -1,55 +1,8 @@
 use crate::{HydrationCtx, IntoView};
 use cfg_if::cfg_if;
 use leptos_reactive::{signal_prelude::*, use_context, RwSignal};
-use server_fn::{ServerFnError, ServerFnErrorErr};
-use std::{borrow::Cow, collections::HashMap, error, fmt, ops, sync::Arc};
-
-/// This is a result type into which any error can be converted,
-/// and which can be used directly in your `view`.
-///
-/// All errors will be stored as [`Error`].
-pub type Result<T, E = Error> = core::result::Result<T, E>;
-
-/// A generic wrapper for any error.
-#[derive(Debug, Clone)]
-#[repr(transparent)]
-pub struct Error(Arc<dyn error::Error + Send + Sync>);
-
-impl Error {
-    /// Converts the wrapper into the inner reference-counted error.
-    pub fn into_inner(self) -> Arc<dyn error::Error + Send + Sync> {
-        Arc::clone(&self.0)
-    }
-}
-
-impl ops::Deref for Error {
-    type Target = Arc<dyn error::Error + Send + Sync>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl fmt::Display for Error {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-impl<T> From<T> for Error
-where
-    T: std::error::Error + Send + Sync + 'static,
-{
-    fn from(value: T) -> Self {
-        Error(Arc::new(value))
-    }
-}
-
-impl From<ServerFnError> for Error {
-    fn from(e: ServerFnError) -> Self {
-        Error(Arc::new(ServerFnErrorErr::from(e)))
-    }
-}
+use server_fn::error::{Error};
+use std::{borrow::Cow, collections::HashMap};
 
 /// A struct to hold all the possible errors that could be provided by child Views
 #[derive(Debug, Clone, Default)]
